refactor(js): split remaining functions into ES modules
Steps:
- Create `constants/api.js` with `urlAPI`
- Create `constants/dom.js` with `appButton` and `appContent`
- Create `services/api.js` with `apiAddLink` and `apiAjaxHandler`
- Create `components/message.js` with `messageAdd`, `messageRemove`, `messageError404Add` and `messageAlertAdd`
- Create `components/filter.js` with `filterSetAction`, `filterActive` and `filterRemoveContent`
- Create `components/search.js` with `searchAdd` and `searchRemove`
- Create `components/pagination.js` with `paginationAdd` and `paginationRemove`
- Create `components/adblock.js` with `adblockVerify`
- Reduce `scripts.js` to entry point with imports and initialization logic (750 → 66 lines)

Reason:
- Part of the modularization effort to split `scripts.js` into separate ES modules
- Resolves `max-lines` warning (no file exceeds 500 lines)

References:
- Issue #104
