+++ assets/js/constants/api.js
@@ -1,0 +1,7 @@
+/**
+ * @const urlAPI
+ * @description API route 'Rick and morty'
+ * @type {String}
+ * @see Used in: {@link apiAjaxHandler}
+ */
+export const urlAPI = "https://rickandmortyapi.com/api/";
