+++ assets/js/constants/dom.js
@@ -1,0 +1,15 @@
+/**
+ * @var appButton
+ * @description App button
+ * @type {HTMLElement}
+ * @see Used in: {@link click}
+ */
+export const appButton = document.getElementById("appButton");
+
+/**
+ * @var appContent
+ * @description App content
+ * @type {HTMLElement}
+ * @see Used in: {@link filterSetAction}, {@link apiAddLink}, {@link filterRemoveContent}, {@link searchCreate}, {@link searchRemove}, {@link paginationCreate}, {@link paginationRemove}, {@link click}
+ */
+export const appContent = document.getElementById("appContent");
