+++ assets/js/scripts.js
@@ -5,699 +5,15 @@
  * @see {@link https://beatrizsmerino.github.io/rick-and-morty/}
  */
 
-import { delay } from "./utils/timer.js";
+import { urlAPI } from "./constants/api.js";
+import { appButton, appContent } from "./constants/dom.js";
 import { svgMe } from "./utils/svg.js";
+import { apiAddLink, apiAjaxHandler } from "./services/api.js";
+import { adblockVerify } from "./components/adblock.js";
 import { addCurrentYear } from "./components/copyright.js";
-import { loaderAdd, loaderRemove } from "./components/loader.js";
-import { cardCreate, cardMoveImage, cardWhenClicked } from "./components/card.js";
-
-/**
- * @const urlAPI
- * @description API route 'Rick and morty'
- * @type {String}
- * @see Used in: {@link apiAjaxHandler}
- */
-const urlAPI = "https://rickandmortyapi.com/api/";
-
-/**
- * @var appButton
- * @description App button
- * @type {HTMLElement}
- * @see Used in: {@link click}
- */
-const appButton = document.getElementById("appButton");
-
-/**
- * @var appContent
- * @description App content
- * @type {HTMLElement}
- * @see Used in: {@link filterSetAction}, {@link apiAddLink}, {@link filterRemoveContent}, {@link searchCreate}, {@link searchRemove}, {@link paginationCreate}, {@link paginationRemove}, {@link click}
- */
-const appContent = document.getElementById("appContent");
-
-
-
-
-
-// API
-// ////////////////////////////////
-
-/**
- * @function apiAddLink
- * @description Add link of the API to the app content
- * @param {String} url - root of the API
- * @see Used in: {@link click}
- */
-function apiAddLink(url) {
-	const linkId = document.getElementById("linkApi");
-
-	if (!linkId) {
-		const link = document.createElement("a");
-		const linkText = document.createTextNode(url);
-
-		link.setAttribute("id", "linkApi");
-		link.setAttribute("href", url);
-		link.setAttribute("target", "_blank");
-		link.appendChild(linkText);
-		appContent.appendChild(link);
-	}
-}
-
-/**
- * @function apiAjaxHandler
- * @description API request
- * @param {String} url - root of the API
- * @param {String} action - name of the action to excute
- * @return {object}
- * @see Used inside: {@link loaderAdd}, {@link loaderRemove}, {@link filterSetAction}
- * @see Used in: {@link searchAdd}, {@link paginationAdd}, {@link click}
- */
-function apiAjaxHandler(url, action) {
-	loaderAdd();
-
-	fetch(url).
-		then(handleResponse).
-		then(function(data) {
-			// console.log("%c--- Promise 2 ---", "padding: 0.5rem 1rem; color: #C0C0C0; background-color: #454545;");
-			// console.info(data);
-			// console.info('data is', data);
-			const timer = setInterval(function() {
-				clearInterval(timer);
-				loaderRemove();
-				messageRemove("messageError404");
-				filterSetAction(action, appContent, data);
-			}, 3000);
-		}).
-		catch(function(error) {
-			// console.warn('error is', error);
-			if (error.status === 404) {
-				loaderRemove();
-				messageError404Add();
-			}
-		});
-
-	function handleResponse(response) {
-		const contentType = response.headers.get("content-type");
-		if (contentType.includes("application/json")) {
-			return handleJSONResponse(response);
-		} else if (contentType.includes("text/html")) {
-			return handleTextResponse(response);
-		}
-
-		// Other response types as necessary. I haven't found a need for them yet though.
-		throw new Error(`Sorry, content-type ${contentType} not supported`);
-	}
-
-	function handleJSONResponse(response) {
-		return response.json().then(json => {
-			if (response.ok) {
-				return json;
-			}
-
-			return Promise.reject(Object.assign(new Error(response.statusText), {
-				...json,
-				"status": response.status,
-			}));
-		});
-	}
-
-	function handleTextResponse(response) {
-		return response.text().then(text => {
-			if (response.ok) {
-				return text;
-			}
-
-			return Promise.reject(Object.assign(new Error(response.statusText), {
-				"status": response.status,
-				"err": text,
-			}));
-		});
-	}
-}
-
-
-
-// FILTER
-// ////////////////////////////////
-
-/**
- * @function filterSetAction
- * @description List of functions to choose from
- * @param {String} action - name of the action to excute
- * @param {HTMLElement} elementDom - DOM element where the response data is inserted
- * @param {Object} dataResponse - response data of the ajax handler (json)
- * @see Used inside: {@link filterAdd}, {@link filterAddContent}
- * @see Used in: {@link apiAjaxHandler}
- */
-function filterSetAction(action, elementDom, dataResponse) {
-	if (action === "filterAdd") {
-		filterAdd(elementDom, dataResponse);
-	} else if (action === "filterAddContent") {
-		filterAddContent(elementDom, dataResponse);
-	}
-}
-
-/**
- * @function filterAdd
- * @description Add navigation menu filtering through the 3 types of data (characters, locations and episodes) to the app content.
- * @param {HTMLElement} elementDom - DOM element where the filter is inserted
- * @param {Object} responseData - response data of the ajax handler (json)
- * @see Used in: {@link filterSetAction}
- */
-function filterAdd(elementDom, responseData) {
-	const navId = document.getElementById("filter");
-
-	if (!navId) {
-		const list = document.createElement("ul");
-		list.setAttribute("id", "filter");
-		list.setAttribute("class", "filter");
-
-		for (const key in responseData) {
-			if (Object.hasOwn(responseData, key)) {
-				const element = responseData[key];
-				const item = document.createElement("li");
-				const itemText = document.createTextNode(key);
-
-				item.setAttribute("data-filter", key);
-				item.setAttribute("data-url", `${element}?page=1`);
-				item.setAttribute("class", "filter__item");
-
-				item.appendChild(itemText);
-				list.appendChild(item);
-			}
-		}
-
-		elementDom.appendChild(list);
-	}
-}
-
-/**
- * @function filterActive
- * @description Add class 'is-active' to the item of the navigation clicked.
- * @param {HTMLCollectionOf} item - filter list
- * @param {Element} thisActive - filter selected
- * @see Used in: {@link functionAnonimAutoExecuted}
- */
-function filterActive(item, thisActive) {
-	for (let index = 0; index < item.length; index++) {
-		const element = item[index];
-		if (element.classList.contains("is-active")) {
-			// console.log("has class");
-			element.classList.remove("is-active");
-		}
-		thisActive.classList.add("is-active");
-	}
-}
-
-/**
- * @function filterAddContentInfo
- * @description Insert information to the content with the number of results of the request
- * @param {Object} responseData - response data of the ajax handler (json)
- * @return {Element}
- * @see Used in: {@link filterAddContent}
- */
-function filterAddContentInfo(responseData) {
-	// console.table(responseData.info);
-	const listInfo = document.createElement("div");
-	listInfo.setAttribute("class", "list-info");
-	listInfo.innerHTML = `<p><strong>Results: </strong>${responseData.info.count}</p>`;
-
-	return listInfo;
-}
-
-/**
- * @function filterAddContentResults
- * @description Insert results to the content of the request
- * @param {Object} responseData - response data of the ajax handler (json)
- * @return {Element}
- * @see Used inside: {@link cardCreate}, {@link cardMoveImage}, {@link cardWhenClicked}
- * @see Used in: {@link filterAddContent}
- */
-function filterAddContentResults(responseData) {
-	// console.table(responseData.results);
-	const listCards = document.createElement("div");
-	const listCardsInner = document.createElement("div");
-
-	listCards.setAttribute("class", "list-cards");
-
-	listCardsInner.setAttribute("id", "filterResult");
-	listCardsInner.setAttribute("class", "list-cards__inner");
-
-	cardCreate(listCardsInner, responseData);
-	cardMoveImage();
-	cardWhenClicked();
-
-	listCards.appendChild(listCardsInner);
-
-	return listCards;
-}
-
-/**
- * @function filterAddContent
- * @description Add the filter content application
- * @param {HTMLElement} elementDom - DOM element where the filter content is inserted
- * @param {Object} responseData - response data of the ajax handler (json)
- * @see Used inside: {@link filterAddContentInfo}, {@link filterAddContentResults}...
- * @see Used in: {@link filterSetAction}
- */
-function filterAddContent(elementDom, responseData) {
-	const list = document.createElement("section");
-	list.setAttribute("class", "list");
-	list.setAttribute("id", "list");
-
-	const infoContent = filterAddContentInfo(responseData);
-	const resultsContent = filterAddContentResults(responseData);
-
-	/**
-	 * @function filterAddAllContent
-	 * @description Insert the content
-	 * @see {@link filterFoundContent}
-	 */
-	function filterAddAllContent() {
-		list.appendChild(infoContent);
-		list.appendChild(resultsContent);
-		elementDom.appendChild(list);
-	}
-
-	/**
-	 * @function filterFoundContent
-	 * @description Check if the contents of one filter are shown and before loading another one, delete it
-	 * @see Used inside: {@link filterAddAllContent}, {@link filterRemoveContent}, {@link paginationRemove}, {@link paginationAdd}
-	 */
-	function filterFoundContent() {
-		const element = document.querySelectorAll(".list");
-		if (typeof element !== "undefined") {
-			// console.dir(element);
-			// console.log(element.length);
-			if (element.length == 0) {
-				filterAddAllContent();
-				paginationAdd(responseData);
-
-				return false;
-			}
-			filterRemoveContent();
-			paginationRemove();
-			filterAddAllContent();
-			paginationAdd(responseData);
-
-			return true;
-		}
-
-		return false;
-	}
-
-	const timer = setInterval(function() {
-		if (filterFoundContent()) {
-			clearInterval(timer);
-		}
-	}, 100);
-}
-
-/**
- * @function filterRemoveContent
- * @description Remove the selected filter content of the application content
- * @see Used in: {@link searchAdd}, {@link filterAddContent}, {@link paginationAdd}, {@link functionAnonimAutoExecuted}
- */
-function filterRemoveContent() {
-	const list = document.querySelectorAll(".list");
-	for (let index = 0; index < list.length; index++) {
-		const element = list[index];
-		if (element && element.innerHTML !== "") {
-			appContent.removeChild(element);
-		}
-	}
-}
-
-// MESSAGE
-// ////////////////////////////////
-/**
- * @function messageAdd
- * @description Create the message component
- * @param {String} messageId - Id for events js
- * @param {String} messageClass - Class css with modifier BEM of the message
- * @param {String} messageText - Text of the message
- * @see Used inside: {@link messageCloseAdd}
- * @see Used in: {@link messageError404Add}, {@link messageAlertAdd}
- */
-function messageAdd(messageId, messageClass, messageText) {
-	const messageDom = document.getElementById(messageId);
-
-	let messageButton = "";
-	switch (messageClass) {
-	case "message-alert":
-		messageButton = messageCloseAdd();
-		messageButton = messageButton.outerHTML;
-		break;
-	case "message-error404":
-		messageButton = "";
-		break;
-	default:
-		break;
-	}
-
-	if (!messageDom) {
-		const messageElem = document.createElement("div");
-		messageElem.setAttribute("id", messageId);
-		messageElem.setAttribute("class", `message ${messageClass}`);
-
-		const templateMessage = `
-			<div class="message__inner">
-				${messageButton}
-				<div class="message__content">
-					<p>
-						${messageText}
-					</p>
-				</div>
-			</div>
-		`;
-		messageElem.innerHTML = templateMessage;
-
-		return messageElem;
-	}
-
-	return null;
-}
-
-/**
- * @function messageCloseAdd
- * @description Create a button for remove the message
- * @see Used in: {@link messageAlertAdd}
- */
-function messageCloseAdd() {
-	const buttonDom = document.createElement("button");
-	buttonDom.classList.add("message__button-close");
-
-	return buttonDom;
-}
-
-/**
- * @function messageRemove
- * @description Remove the message component
- * @param {String} messageId - id of the message to remove
- * @see Used in: {@link apiAjaxHandler}
- */
-function messageRemove(messageId) {
-	const messsage = document.getElementById(messageId);
-	if (messsage) {
-		messsage.parentElement.removeChild(messsage);
-
-		// console.log("Message removed");
-	}
-}
-
-/**
- * @function messageError404Add
- * @description Add message error 404 (search not found)
- * @see Used inside: {@link messageAdd}
- * @see Used in: {@link apiAjaxHandler}
- */
-function messageError404Add() {
-	filterRemoveContent();
-	paginationRemove();
-
-	const templateMessage = messageAdd("messageError404", "message-error404", "Error 404.</br> Search not found");
-	appContent.appendChild(templateMessage);
-}
-
-/**
- * @function messageAlertAdd
- * @description Create a alert personalized
- * @param {String} messageName - Name of message in camellCase
- * @param {String} messageText - Text of message
- * @see Used inside: {@link messageAdd}, {@link messageRemove}
- */
-function messageAlertAdd(messageName, messageText) {
-	const messageId = `messageAlert${messageName}`;
-	const templateMessage = messageAdd(messageId, "message-alert", messageText);
-	document.body.appendChild(templateMessage);
-	document.querySelector(`#${messageId} .message__button-close`).addEventListener("click", function() {
-		messageRemove(messageId);
-	});
-}
-
-
-// SEARCH
-// ////////////////////////////////
-
-/**
- * @function searchCreate
- * @description Create searcher
- * @see Used in: {@link searchAdd}
- */
-function searchCreate() {
-	const searchDom = document.createElement("div");
-	const searchInnerDom = document.createElement("div");
-	const searchIconDom = document.createElement("div");
-	const searchInput = document.createElement("input");
-
-	searchDom.setAttribute("id", "search");
-	searchDom.setAttribute("class", "search");
-	searchInnerDom.setAttribute("class", "search__inner");
-	searchIconDom.setAttribute("class", "search__icon icon-magnifying-glass");
-
-	searchInput.setAttribute("id", "searchInput");
-	searchInput.setAttribute("class", "search__input");
-
-	searchInnerDom.appendChild(searchIconDom);
-	searchInnerDom.appendChild(searchInput);
-	searchDom.appendChild(searchInnerDom);
-	appContent.appendChild(searchDom);
-}
-
-/**
- * @function searchGet
- * @description Get the active filter to find it.
- * @param {Element} filterSelected - filter selected
- * @see Used in: {@link searchAdd}
- */
-function searchGet(filterSelected) {
-	const searchInput = document.getElementById("searchInput");
-	const filterActiveText = filterSelected.getAttribute("data-filter");
-	searchInput.setAttribute("placeholder", `Search by name of ${filterActiveText}`);
-
-	let searchBy = "";
-	switch (filterActiveText) {
-	case "characters":
-		searchBy = "character";
-		break;
-	case "episodies":
-		searchBy = "episode";
-		break;
-	case "locations":
-		searchBy = "location";
-		break;
-	default:
-		break;
-	}
-
-	// console.assert(searchBy !== "", "Not Search");
-	// console.log(searchBy);
-
-	return searchBy;
-}
-
-/**
- * @function searchAdd
- * @description Add searcher
- * @param {Element} filterSelected - filter selected
- * @see Used inside: {@link searchCreate}, {@link searchGet}...
- * @see Used in: {@link functionAnonimAutoExecuted}
- */
-function searchAdd(filterSelected) {
-	searchCreate();
-
-	const searchBy = searchGet(filterSelected);
-
-	/**
-	 * @description Search by selected filter name when typing in the search engine input.
-	 * @event keyup
-	 * @type {Object}
-	 * @see Used inside: {@link delay}, {@link filterRemoveContent}, {@link paginationRemove}, {@link apiAjaxHandler}
-	 */
-	document.getElementById("searchInput").addEventListener(
-		"keyup",
-		delay(function() {
-			const valueInput = this.value;
-
-			filterRemoveContent();
-			paginationRemove();
-			apiAjaxHandler(`${urlAPI + searchBy}/?name=${valueInput}`, "filterAddContent");
-
-			// console.log(this);
-			// console.log(this.value);
-			// console.log(urlAPI + searchBy + "/?name=" + valueInput);
-			// console.assert(valueInput, "Input hasn`t value");
-		}, 500),
-	);
-}
-
-/**
- * @function searchRemove
- * @description Remove searcher
- * @see Used in: {@link functionAnonimAutoExecuted}
- */
-function searchRemove() {
-	const search = document.getElementById("search");
-	if (search) {
-		appContent.removeChild(search);
-	}
-}
-
-// PAGINATION
-// ////////////////////////////////
-
-/**
- * @function paginationCreate
- * @description Create pagination
- * @param {Object} responseData - response data of the ajax handler (json)
- * @see Used in: {@link paginationAdd}
- */
-function paginationCreate(responseData) {
-	const pagination = document.createElement("div");
-	pagination.setAttribute("class", "pagination");
-	pagination.setAttribute("id", "pagination");
-
-	const buttonNext = document.createElement("div");
-	buttonNext.setAttribute("class", "pagination__button icon-chevron-right");
-	buttonNext.setAttribute("id", "buttonNext");
-	buttonNext.setAttribute("data-url", responseData.info.next);
-
-	const buttonPrev = document.createElement("div");
-	buttonPrev.setAttribute("class", "pagination__button icon-chevron-left");
-	buttonPrev.setAttribute("id", "buttonPrev");
-	buttonPrev.setAttribute("data-url", responseData.info.prev);
-
-	pagination.appendChild(buttonPrev);
-	pagination.appendChild(buttonNext);
-	appContent.appendChild(pagination);
-}
-
-/**
- * @function paginationSetCounter
- * @description Create counter pagination
- * @param {Object} responseData - response data of the ajax handler (json)
- * @see Used in: {@link paginationAdd}
- */
-function paginationSetCounter(responseData) {
-	const buttonPrev = document.getElementById("buttonPrev");
-
-	const paginationTotal = responseData.info.pages;
-	const paginationNext = responseData.info.next;
-	let paginationNow = 0;
-
-	// console.log(paginationNext);
-	if (paginationNext == "") {
-		paginationNow = paginationTotal;
-	} else {
-		const paginationNextUrl = new URL(paginationNext);
-		const numPageNext = paginationNextUrl.searchParams.get("page");
-		paginationNow = parseInt(numPageNext) - 1;
-
-		if (paginationNow <= 1 || isNaN(paginationNow)) {
-			paginationNow = 1;
-		}
-	}
-
-	// console.log(paginationNext);
-	// console.log(paginationNow);
-
-	const paginationCounter = document.createElement("div");
-	const paginationCounterText = document.createTextNode(`${paginationNow.toString()} - ${paginationTotal.toString()}`);
-
-	paginationCounter.setAttribute("class", "pagination__counter");
-	paginationCounter.appendChild(paginationCounterText);
-
-	// console.log(paginationCounter);
-
-	buttonPrev.parentNode.insertBefore(paginationCounter, buttonPrev.nextSibling);
-}
-
-/**
- * @function paginationAdd
- * @description Add pagination
- * @param {Object} responseData - response data of the ajax handler (json)
- * @see Used inside: {@link paginationCreate}, {@link paginationSetCounter}...
- * @see Used in: {@link filterFoundContent}
- */
-function paginationAdd(responseData) {
-	paginationCreate(responseData);
-	paginationSetCounter(responseData);
-
-	const button = document.getElementsByClassName("pagination__button");
-	for (let index = 0; index < button.length; index++) {
-		const element = button[index];
-
-		/**
-		 * @description Remove/Add content and pagination by selecting the filter from the navigation menu.
-		 * @event click
-		 * @type {object}
-		 * @see Used inside: {@link filterRemoveContent}, {@link paginationRemove}, {@link apiAjaxHandler}
-		 */
-		element.addEventListener("click", function() {
-			const url = this.getAttribute("data-url");
-
-			if (url !== "") {
-				filterRemoveContent();
-				paginationRemove();
-				apiAjaxHandler(url, "filterAddContent");
-			}
-		});
-	}
-}
-
-/**
- * @function paginationRemove
- * @description Remove pagination
- * @see Used in: {@link functionAnonimAutoExecuted}, {@link filterAddContent}, {@link searchAdd}, {@link paginationAdd}
- */
-function paginationRemove() {
-	const pagination = document.getElementById("pagination");
-	if (pagination) {
-		appContent.removeChild(pagination);
-	}
-}
-
-// ADBLOCK
-// ////////////////////////////////
-
-//
-// @function adblockDetected
-// @description Callback executed if adblock is installed
-// @see Used inside {@link messageAlert}
-// @see Used in: {@link checkAdblock}
-//
-function adblockDetected() {
-	const message = "<i class='message-alert__icon icon-warning'></i> AdBlock is enabled";
-
-	// console.warn(message);
-	// alert(message);
-	messageAlertAdd("AdBlock", message);
-}
-
-//
-// @function adblockDisabled
-// @description Callback executed if adblock is disabled
-// @see Used in: {@link checkAdblock}
-//
-function adblockDisabled() {
-	// let message = "AdBlock is not enabled";
-	// console.info(message);
-}
-
-//
-// @function adblockVerify
-// @description Verify if the user has installed the Adblock browser extension
-// @see Used inside: {@link adblockDetected} {@link adblockDisabled}
-// @see Used in: {@link functionAnonimAutoExecuted}
-//
-function adblockVerify() {
-	if (typeof blockAdBlock === "undefined") {
-		adblockDetected();
-	} else {
-		blockAdBlock.onDetected(adblockDetected);
-		blockAdBlock.onNotDetected(adblockDisabled);
-	}
-}
+import { filterActive, filterRemoveContent } from "./components/filter.js";
+import { paginationRemove } from "./components/pagination.js";
+import { searchAdd, searchRemove } from "./components/search.js";
 
 /**
  * @description Get API data
